==> 1_cleaning_top_songs.ipynb <==
# CLEANING TOP SONGS

Source: https://www.kaggle.com/edumucelli/spotifys-worldwide-daily-song-ranking/

## Import things

import pandas as pd
import numpy as np

## Load the data

top_songs = pd.read_csv('./data.csv')

## Get to know the data & some initial fix ##

top_songs.head()

top_songs.shape

top_songs.info()

COMMENT: I think that the data over all looks good. I need to do some adjustments, based on what I'm goign to use the data for. I might also get rid of some data that I am not going to use.

#### Look at the column Position

top_songs['Position'].isnull().sum()

top_songs['Position'].unique()

top_songs['Position'].dtype

COMMENT: Nothing really surprising here

#### Look at the column Track Name

top_songs['Track Name'].isnull().sum()

top_songs['Track Name'].unique()

#### Show rows that has NULL values

top_songs[top_songs['Track Name'].isnull()]

# look at unique regions that are missing data
top_songs[top_songs['Track Name'].isnull()]['Region'].unique()
# this code might be wrong...

I think I have to drop the rows with NULL values for Trackname and Artist.
There is no way I can in any way fill this in. And it is going to be really hard to put together with the analysis from the lyrics, since I do not know what song it is. 
I is 657 songs out of 3,441,197. That is 0.019%...so I still have way more than 95% of my data.
The missing data seem to be distributed over different regions and not only one or a few.

DROP THE ROWS WITH NULL VALUES IN THE END...some rows might be excluded by other reasons.

top_songs['Track Name'].dtype

COMMENT: I think I might have a need to exclude asia as a region du to different language signs. Will be really hard for the lyrics part. Might have to exclude other regions as well, based on that.

#### Look at the column Artist

top_songs['Artist'].isnull().sum()

top_songs['Artist'].unique()

top_songs['Artist'].dtype

COMMENT: I think I might have a need to exclude asia as a region du to different language signs. Will be really hard for the lyrics part. Might have to exclude other regions as well, based on that.

#### Look at the column Streams

top_songs['Streams'].isnull().sum()

top_songs['Streams'].dtype

COMMENT: Looks good. No NULL-values and they are all integer. Hurray!

#### Look at the column Date

top_songs['Date'].isnull().sum()

top_songs['Date'].dtype

# converting into datetime-format
top_songs['Date'] = pd.to_datetime(top_songs['Date'])

top_songs['Date'].dtype

top_songs['Date'].min()

top_songs['Date'].max()

# I do want to remove the 9 days in 2018. I only want data for 2017. That is one of my limitations. 

# make new column for year. Then drop all rows from 2018
top_songs['Year'] = top_songs['Date'].dt.year

top_songs.head()

#  make new data frame with only data for year 2017
top_songs = top_songs[top_songs['Year'] != 2018]

top_songs['Date'].max()

# make new column for month. 
top_songs['Month'] = top_songs['Date'].dt.month

# make new column for day. 
top_songs['Day'] = top_songs['Date'].dt.day

top_songs.head()

COMMENT: When all the data put together is in one data frame the Date column has to be transformed into datetime-format again (that format is not saved I think i the csv file). I might also like to make separat columns for year, month and day separatly. 

#### Look at the column Region (will be renamed to Country)

# Rename the column Region to Country
top_songs.rename(columns={'Region': 'Country'}, inplace=True)

top_songs.columns

top_songs['Country'].isnull().sum()

top_songs['Country'].unique()

np.sort(top_songs['Country'].unique())

top_songs['Country'].dtype

On Kaggle someone already created a dictonary with the real names of the regions (countries). Very convenient!<BR />
I will put it in the read me file under info about data or something similar<BR />
https://www.kaggle.com/mrncst/simple-exploratory-analysis-of-spotify-songs<BR />
I do not want to change my names however (a short version is esier to work with).<BR />
Why do I need to know this. Bc I do want to group countries into bigger "regions" based on continents.

'ar':'Argentina'<BR />
'at':'Austria'<BR />
'au':'Australia'<BR />
'be':'Belgium'<BR />
'bo':'Bolivia'<BR />
'br':'Brazil'<BR />
'ca':'Canada'<BR />
'ch':'Switzerland'<BR />
'cl':'Chile'<BR />
'co':'Columbia'<BR />
'cr':'CostaRica'<BR />
'cz':'CzechRepublic'<BR />
'de':'Germany'<BR />
'dk':'Denmark'<BR />
'do':'DominicanRepublic'<BR />
'ec':'Ecuador'<BR />
'ee':'Estonia'<BR />
'es':'Spain'<BR />
'fi':'Finland'<BR />
'fr':'France'<BR />
'gb':'UnitedKingdom'<BR />
'global':'World'<BR />
'gr':'Greece'<BR />
'gt':'Guatemala'<BR />
'hk':'HongKong'<BR />
'hn':'Honduras'<BR />
'hu':'Hungary'<BR />
'id':'Indonesia'<BR />
'ie':'Ireland'<BR />
'is':'Iceland'<BR />
'it':'Italy'<BR />
'jp':'Japan'<BR />
'lt':'Lithuania'<BR />
'lu':'Luxemborg'<BR />
'lv':'Latvia'<BR />
'mx':'Mexico',<BR /> 
'my':'Malaysia'<BR />
'nl':'Netherlands'<BR />
'no':'Norway'<BR />
'nz':'NewZealand'<BR />
'pa':'Panama'<BR />
'pe':'Peru',<BR />
'ph':'Philippines'<BR />
'pl':'Poland'<BR />
'pt':'Portugal'<BR />
'py':'Paraguay'<BR />
'se':'Sweden'<BR />
'sg':'Singapore'<BR />
'sk':'Slovakia'<BR />
'sv':'ElSalvador'<BR />
'tr':'Turkey'<BR />
'tw':'Taiwan'<BR />
'us':'USA'<BR />
'uy':'Uruguay'<BR />

# GROUP THE CONTRIES INTO REGIONS. Make a new column that indicated what bigger region (based on continent) 
# a country in the column Country belongs to. 
# Name the "new column" Region

**South America (8)**<BR />
'ar':'Argentina'<BR />
'bo':'Bolivia'<BR />
'br':'Brazil'<BR />
'co':'Columbia'<BR />
'ec':'Ecuador'<BR />
'pe':'Peru',<BR />
'py':'Paraguay'<BR />
'uy':'Uruguay'<BR />

**Europe (26)**<BR />
'at':'Austria'<BR />
'be':'Belgium'<BR />
'ch':'Switzerland'<BR />
'cz':'CzechRepublic'<BR />
'de':'Germany'<BR />
'dk':'Denmark'<BR />
'ee':'Estonia'<BR />
'es':'Spain'<BR />
'fi':'Finland'<BR />
'fr':'France'<BR />
'gb':'UnitedKingdom'<BR />
'gr':'Greece'<BR />
'hu':'Hungary'<BR />
'ie':'Ireland'<BR />
'is':'Iceland'<BR />
'it':'Italy'<BR />
'lt':'Lithuania'<BR />
'lu':'Luxemborg'<BR />
'lv':'Latvia'<BR />
'nl':'Netherlands'<BR />
'no':'Norway'<BR />
'pl':'Poland'<BR />
'pt':'Portugal'<BR />
'se':'Sweden'<BR />
'sk':'Slovakia'<BR />
'tr':'Turkey'<BR />

**Oceania (2)**<BR />
'au':'Australia'<BR />
'nz':'NewZealand'<BR />


**Nothe America & Caribean (9)**<BR />
'ca':'Canada'<BR />
'cr':'CostaRica'<BR />
'do':'DominicanRepublic'<BR />
'gt':'Guatemala'<BR />
'hn':'Honduras'<BR />
'mx':'Mexico',<BR /> 
'pa':'Panama'<BR />
'sv':'ElSalvador'<BR />
'us':'USA'<BR />

**Asia (7)**<BR />
'hk':'HongKong'<BR />
'id':'Indonesia'<BR />
'jp':'Japan'<BR />
'my':'Malaysia'<BR />
'ph':'Philippines'<BR />
'sg':'Singapore'<BR />
'tw':'Taiwan'<BR />

**Other (1)**<BR />
'global':'World'<BR />

# Create new column for Region (based on continent)
def country_to_region(country):
    sa = ['ar', 'bo', 'br', 'co', 'ec', 'pe', 'py', 'uy']
    eu = ['at', 'be', 'ch', 'cz', 'de', 'dk', 'ee', 'es', 'fi', 'fr', 
          'gb', 'gr', 'hu', 'ie', 'is', 'it', 'lt', 'lu', 'lv', 'nl', 
          'no', 'pl', 'pt', 'se', 'sk', 'tr']
    oc = ['au', 'nz' ]
    na = ['ca', 'cr', 'do', 'gt', 'hn', 'mx', 'pa', 'sv', 'us']
    asi = ['hk', 'id', 'jp', 'my', 'ph', 'sg', 'tw']
    
    if country in sa:
        return 'sa'
    elif country in eu:
        return 'eu'
    elif country in oc:
        return 'oc'
    elif country in na:
        return 'na'
    elif country in asi:
        return 'as'
    else:
        return 'world'

top_songs['Region'] = top_songs['Country'].apply(country_to_region)

top_songs['Region'].unique()

COMMENT: I might want to focus more on one of the regions (based on continent). The column Global is what I think it it, it can be used for feature engineering. You might want to see if a country is deviating a lot from global/average for example.

**Look at missing value again for the columns Track Name and Artist**

top_songs['Track Name'].isnull().sum()

top_songs['Artist'].isnull().sum()

top_songs.shape

Drop rows with missing values in any column (we know that there are only missing values in the columns Track Name and Artist)

top_songs = top_songs.dropna(axis=0, how='any')

top_songs.shape

# Look that we have successfully dropped 657 rows. (It also tells us that all the missing values that existed in the 
# column Track Name corresponded ti missing values in the column Artist.)
3356486-3355829

# Keep in mind that some positions will be missing after dropping rows

Change the order of the columns

top_songs.head()

top_songs = top_songs[['Position', 'Track Name', 'Artist', 'Streams', 'URL', 'Date', 'Year', 'Month', 'Day', 'Country', 'Region']]

top_songs.head()

## Save to a csv

# load the data to a .csv
top_songs.to_csv('top_songs_clean.csv')
