==> 2_cleaning_lyrics.ipynb <==
# CLEANING LYRICS #

Source: https://www.kaggle.com/artimous/every-song-you-have-heard-almost

## Import things

import pandas as pd
import numpy as np

## Load the data

artist = pd.read_csv('./ArtistUrl.csv')
lyrics1 = pd.read_csv('./Lyrics1.csv')
lyrics2 = pd.read_csv('./Lyrics2.csv')

## Get to know the data

artist.head()

lyrics1.head()

lyrics1.shape

lyrics1.isnull().sum()

lyrics2.head()

lyrics2.shape

lyrics2.isnull().sum()

### Make the 2 data frames lyrics 1 and 2 into one 

lyrics = pd.concat([lyrics1, lyrics2])

lyrics.shape

lyrics.head()

lyrics.isnull().sum()

#### Drop rows where Song or Lyrics are missing

I made this decision bc it was only 8 rows and imputation in this case was not a good option.

lyrics = lyrics.dropna(axis=0, how='any')

lyrics.shape

#### Clean the lyrics from strange stuff

I did some inital cleaning, need to do more when everything is in one data frame.

# \r\n = new line
# replace '\r\n', '\n\n', '\r', '\n' with ' '
lyrics['Lyrics'] = lyrics['Lyrics'].str.replace('\r\n', ' ')
lyrics['Lyrics'] = lyrics['Lyrics'].str.replace('\n\n', ' ')
lyrics['Lyrics'] = lyrics['Lyrics'].str.replace('\r', ' ')
lyrics['Lyrics'] = lyrics['Lyrics'].str.replace('\n', ' ')

lyrics['Lyrics'].head(15)

#### Convert to lower case letters

lyrics['Lyrics'] = lyrics['Lyrics'].str.lower()

lyrics['Lyrics'].head(15)

#### Look and see if Chantaje are in the list

I just tried looking for one song to see if it was in the data frame.

lyrics.loc[lyrics['Song'] == 'Chantaje']

Since the lyrics was last updated areound last summer all songlyrics are not going to be in the lyrics.
I need to find additional scources for to get more song lyrics.

## Save to a csv

# load the data to a .csv
lyrics.to_csv('lyrics_clean.csv')
