==> 3_combined_data.ipynb <==
# Combinding data to one Data Frame

## Import things

import pandas as pd
import numpy as np

## Load the data|

songs = pd.read_csv('./top_songs_clean.csv')

lyrics = pd.read_csv('./lyrics_clean.csv')

## Get to know the data

### Songs (Daily top-list data from Spotify)

songs.head(3)

songs.shape

#### Make a column with both Artist name and Track name

I did this since there can be songs with exactly the same name. I also needed a column in common for the 2 data frames to be able to join them together in a good way.

songs['art_song'] = songs['Artist'] + ' - ' + songs['Track Name']

songs.head(3)

### Lyrics (lyrics data from lyrics.com)

lyrics.head(3)

lyrics.shape

#### Make a column with both Band name and Song name

I did this since there can be songs with exactly the same name.  I also needed a column in common for the 2 data frames to be able to join them together in a good way.

lyrics['art_song'] = lyrics['Band'] + ' - ' + lyrics['Song']

lyrics.head(3)

#### Drop duplicated rows based on the new column

lyrics2 = lyrics.drop_duplicates(subset='art_song')

lyrics2.shape

### Make one data frame out of the 2 (songs and lyrics)

df = songs.merge(lyrics2, how='left', on='art_song', indicator=True, sort=True)

df.shape

df.head(3)

## Save the data frame as a CSV-file

# load the data to a csv-file
df.to_csv('songs_lyrics.csv')
