==> 4_spotify_api.ipynb <==
# Getting data from Spotify API

https://developer.spotify.com/web-api/

## Import things

import spotipy
from spotipy.oauth2 import SpotifyClientCredentials

import pandas as pd
import numpy as np

## Creat API client with the right keys

client_credentials_manager = SpotifyClientCredentials('0bc389e00d214082bf48acca68452036', '9b9ad7bee99b490cbc657a388de74dd0')
spotify = spotipy.Spotify(client_credentials_manager=client_credentials_manager)

## Load the data

songs_lyrics = pd.read_csv('./songs_lyrics.csv')

songs_lyrics.head(3)

### Get all unique song uri's

uris = songs_lyrics['URL'].str.replace('https://open.spotify.com/track/', 'spotify:track:')

uris.shape

uris2 = uris.drop_duplicates()

uris2.shape

### Make a list of lists of uris (50 each)

https://stackoverflow.com/questions/312443/how-do-you-split-a-list-into-evenly-sized-chunks

l = uris2.values
n = 50

lists = [l[i:i + n] for i in range(0, len(l), n)]

### Use Spotify API to get data for songs

def get_data(lst):
    results = spotify.audio_features(tracks=lst)
    return results

all_results = []

for lst in lists:
    r = get_data(lst)
    all_results = all_results + r

len(all_results)

# Remove non values (3) and make result to a list
all_avail_results = list(filter(None, all_results))

len(all_avail_results)

## Save result as data frame

df = pd.DataFrame(all_avail_results)

df.head()

Explanation for the Spotify API columns<BR />
https://developer.spotify.com/web-api/get-audio-features/

## Save the data as a CSV-file

# save data to csv-file
df.to_csv('spotify_api_data.csv')
