==> 5_combined_data2.ipynb <==
# Combinding data to one Data Frame 2

## Import things

import pandas as pd
import numpy as np

## Load the data

sl = pd.read_csv('./songs_lyrics.csv')

sa = pd.read_csv('./spotify_api_data.csv')

## Look at the data

sl.head(3)

sa.head(3)

### Make new column in the sl data frame containing only song id from the URL

This is done so that I have something to combine the data frames on.

sl['id'] = sl['URL'].str.replace('https://open.spotify.com/track/', '')

sl.head(3)

### Combine the 2 data frames into one

df = sl.merge(sa, how='left', on='id')

df.shape

df.head(3)

df.columns

### Drop some column that I will not need

I did get more data from the Spotidy API than I am going to use. I will only use columns: acousticness, energy, instrumentalness, mode, tempo, valence (target) from the Spotify API in this project.

df = df.drop(['Unnamed: 0_x', 'URL', 'art_song', 'Unnamed: 0_y', 'Band', 'Song', '_merge', 'analysis_url', 'danceability',
         'duration_ms', 'key', 'liveness', 'loudness', 'speechiness', 'time_signature', 'track_href', 'type', 'uri'], axis=1)

df.head(3)

#### Rename some of the columns

df = df.rename(index=str, columns={"id": "ID", "acousticness": "Acousticness", "energy": "Energy",
                            "instrumentalness": "Instrumentalness", "mode": "Mode", "tempo": "Tempo",
                            "valence": "Valence"})

#### Change order of the columns

df = df[['Position', 'Streams', 'Track Name', 'Artist', 'ID', 'Date', 'Year', 'Month', 'Day', 'Country', 'Region', 'Lyrics',
    'Acousticness', 'Energy', 'Instrumentalness', 'Mode', 'Tempo', 'Valence']]

#### Look at the info about the data to see data types for the columns

df.info()

#### Drop duplicated rows based on the ID-column

df2 = df.drop_duplicates(subset='ID')

#### Look at unique songs

# Number of unique songs
df2.shape

# Number of unique songs where I have lyrics
df2['Lyrics'].notnull().sum()

# Share of songs with lyrics
2502/21372

I only have lyrics for 11.77% of my songs.<BR />
That is a problem. Need to find more lyrics.

## Save the data as a CSV-file

# save data to csv-file
df.to_csv('songs_lyrics_api_data.csv')
