==> 7_combined_data3.ipynb <==
# Combinding data to one Data Frame 3

## Import things

import pandas as pd
import numpy as np

## Load the data

data = pd.read_csv('./songs_lyrics_api_data.csv')

genius = pd.read_csv('./genius_unique.csv')

## Look at the data

data.head()

genius.head()

data.shape

genius.shape

## Combine the 2 data frames

The Genius data only contain lyrics for the ID where the lyrics was Nan in the Data data frame. 
Therefore a left merge om ID is possible.

df = data.merge(genius, how='left', on='ID', indicator=True)

df['_merge'].value_counts()

### Work the Lyrics-columns

df['Lyrics_x'].isnull().sum()

df['Lyrics_x'].fillna(df['Lyrics_y']).isnull().sum()

#### Combine Lyrics_x and Lyrics_y column into one column

df['Lyrics_x'].fillna(df['Lyrics_y'],inplace=True)

df.shape

#### Drop some columns that I do not need

df = df.drop(['Unnamed: 0_x','Unnamed: 0_y'], axis=1)

df.head()

I do not have to read this brutto data frame to CSV. I will make a data fram only containing the 8 markets I'm interested in. Then I can read that to a CSV-file instead.

### Choose only the top 10 music countries (and exclude the 2 asian countries)

df2 = df[(df['Country'] == 'us') | (df['Country'] == 'gb') | (df['Country'] == 'de') | (df['Country'] == 'fr') |
         (df['Country'] == 'ca') | (df['Country'] == 'au') | (df['Country'] == 'it') | (df['Country'] == 'nl')]

df2.head()

#### Drop some columns that I do not need

df2 = df2.drop(['Lyrics_y','_merge'], axis=1)

#### Rename the Lyrics-column

df2 = df2.rename(index=str, columns={"Lyrics_x": "Lyrics"})

df2.head()

## Save the data to a CSV-file

# read top 10 (8) music countries data frame to CSV
df2.to_csv('data_lyrics_top10c.csv')
